Updated /answers docs
